+++ Libplanet.Explorer.Tests/GeneratedBlockChainFixture.cs
@@ -85,7 +85,7 @@ public class GeneratedBlockChainFixture
         IStore store = new MemoryStore();
         var actionEvaluator = new ActionEvaluator(
             _ => policy.BlockAction,
-            new BlockChainStates(store, stateStore),
+            stateStore,
             TypedActionLoader.Create(typeof(SimpleAction).Assembly, typeof(SimpleAction)));
         Block genesisBlock = BlockChain.ProposeGenesisBlock(
             actionEvaluator,