+++ Libplanet.Tests/Action/ActionEvaluatorTest.cs
@@ -86,7 +86,7 @@ namespace Libplanet.Tests.Action
                 transactions: txs).Propose();
             var actionEvaluator = new ActionEvaluator(
                 _ => null,
-                NullChainStates.Instance,
+                new TrieStateStore(new MemoryKeyValueStore()),
                 new SingleActionLoader(typeof(RandomAction)));
             Block stateRootBlock = noStateRootBlock.Sign(
                 GenesisProposer,
@@ -101,10 +101,10 @@ namespace Libplanet.Tests.Action
 
             for (int i = 0; i < repeatCount; ++i)
             {
-                var actionEvaluations = actionEvaluator.Evaluate(noStateRootBlock);
+                var actionEvaluations = actionEvaluator.Evaluate(noStateRootBlock, null);
                 generatedRandomNumbers.Add(
                     (Integer)actionEvaluations[0].OutputState.GetState(txAddress));
-                actionEvaluations = actionEvaluator.Evaluate(stateRootBlock);
+                actionEvaluations = actionEvaluator.Evaluate(stateRootBlock, null);
                 generatedRandomNumbers.Add(
                     (Integer)actionEvaluations[0].OutputState.GetState(txAddress));
             }
@@ -146,7 +146,8 @@ namespace Libplanet.Tests.Action
             Block block = chain.ProposeBlock(miner);
             chain.Append(block, CreateBlockCommit(block));
 
-            var evaluations = chain.ActionEvaluator.Evaluate(chain.Tip);
+            var evaluations = chain.ActionEvaluator.Evaluate(
+                chain.Tip, chain.Store.GetStateRootHash(chain.Tip.PreviousHash));
 
             Assert.False(evaluations[0].InputContext.BlockAction);
             Assert.Single(evaluations);
@@ -180,7 +181,8 @@ namespace Libplanet.Tests.Action
             chain.StageTransaction(tx);
             Block block = chain.ProposeBlock(new PrivateKey());
             chain.Append(block, CreateBlockCommit(block));
-            var evaluations = chain.ActionEvaluator.Evaluate(chain.Tip);
+            var evaluations = chain.ActionEvaluator.Evaluate(
+                chain.Tip, chain.Store.GetStateRootHash(chain.Tip.PreviousHash));
 
             Assert.False(evaluations[0].InputContext.BlockAction);
             Assert.Single(evaluations);
@@ -229,7 +231,7 @@ namespace Libplanet.Tests.Action
                     txHash: BlockContent.DeriveTxHash(txs),
                     lastCommit: null),
                 transactions: txs).Propose();
-            IAccount previousState = actionEvaluator.PrepareInitialDelta(block);
+            IAccount previousState = actionEvaluator.PrepareInitialDelta(genesis.StateRootHash);
 
             Assert.Throws<OutOfMemoryException>(
                 () => actionEvaluator.EvaluateTx(
@@ -237,7 +239,8 @@ namespace Libplanet.Tests.Action
                     tx: tx,
                     previousState: previousState).ToList());
             Assert.Throws<OutOfMemoryException>(
-                () => chain.ActionEvaluator.Evaluate(block).ToList());
+                () => chain.ActionEvaluator.Evaluate(
+                    block, chain.Store.GetStateRootHash(block.PreviousHash)).ToList());
         }
 
         [Fact]
@@ -267,7 +270,7 @@ namespace Libplanet.Tests.Action
             Block genesis = ProposeGenesisBlock(TestUtils.GenesisProposer);
             var actionEvaluator = new ActionEvaluator(
                 policyBlockActionGetter: _ => null,
-                blockChainStates: NullChainStates.Instance,
+                stateStore: new TrieStateStore(new MemoryKeyValueStore()),
                 actionTypeLoader: new SingleActionLoader(typeof(DumbAction)));
 
             Transaction[] block1Txs =
@@ -300,7 +303,7 @@ namespace Libplanet.Tests.Action
                     timestamp: DateTimeOffset.MinValue.AddSeconds(7)),
             };
             foreach ((var tx, var i) in block1Txs.Zip(
-                Enumerable.Range(0, block1Txs.Count()), (x, y) => (x, y)))
+                Enumerable.Range(0, block1Txs.Length), (x, y) => (x, y)))
             {
                 _logger.Debug("{0}[{1}] = {2}", nameof(block1Txs), i, tx.Id);
             }
@@ -309,7 +312,7 @@ namespace Libplanet.Tests.Action
                 genesis,
                 GenesisProposer,
                 block1Txs);
-            IAccount previousState = actionEvaluator.PrepareInitialDelta(block1);
+            IAccount previousState = actionEvaluator.PrepareInitialDelta(null);
             var evals = actionEvaluator.EvaluateBlock(
                 block1,
                 previousState).ToImmutableArray();
@@ -332,7 +335,7 @@ namespace Libplanet.Tests.Action
                 Assert.Equal(expect.Signer, eval.InputContext.Signer);
                 Assert.Equal(GenesisProposer.ToAddress(), eval.InputContext.Miner);
                 Assert.Equal(block1.Index, eval.InputContext.BlockIndex);
-                randomValue = eval.InputContext.Random.Next();
+                randomValue = eval.InputContext.GetRandom().Next();
                 Assert.Equal(
                     (Integer)eval.OutputState.GetState(
                         DumbAction.RandomRecordsAddress),
@@ -343,7 +346,7 @@ namespace Libplanet.Tests.Action
                         .Select(x => x is Text t ? t.Value : null));
             }
 
-            previousState = actionEvaluator.PrepareInitialDelta(block1);
+            previousState = actionEvaluator.PrepareInitialDelta(null);
             ActionEvaluation[] evals1 =
                 actionEvaluator.EvaluateBlock(block1, previousState).ToArray();
             IImmutableDictionary<Address, IValue> dirty1 = evals1.GetDirtyStates();
@@ -412,7 +415,7 @@ namespace Libplanet.Tests.Action
                     timestamp: DateTimeOffset.MinValue.AddSeconds(4)),
             };
             foreach ((var tx, var i) in block2Txs.Zip(
-                Enumerable.Range(0, block2Txs.Count()), (x, y) => (x, y)))
+                Enumerable.Range(0, block2Txs.Length), (x, y) => (x, y)))
             {
                 _logger.Debug("{0}[{1}] = {2}", nameof(block2Txs), i, tx.Id);
             }
@@ -456,7 +459,7 @@ namespace Libplanet.Tests.Action
                 Assert.Equal(block2.Index, eval.InputContext.BlockIndex);
                 Assert.False(eval.InputContext.Rehearsal);
                 Assert.Null(eval.Exception);
-                randomValue = eval.InputContext.Random.Next();
+                randomValue = eval.InputContext.GetRandom().Next();
                 Assert.Equal(
                     eval.OutputState.GetState(
                         DumbAction.RandomRecordsAddress),
@@ -523,12 +526,12 @@ namespace Libplanet.Tests.Action
                 transactions: txs).Propose();
             var actionEvaluator = new ActionEvaluator(
                 policyBlockActionGetter: _ => null,
-                blockChainStates: NullChainStates.Instance,
+                stateStore: new TrieStateStore(new MemoryKeyValueStore()),
                 actionTypeLoader: new SingleActionLoader(typeof(DumbAction)));
 
             DumbAction.RehearsalRecords.Value =
                 ImmutableList<(Address, string)>.Empty;
-            IAccount previousState = actionEvaluator.PrepareInitialDelta(block);
+            IAccount previousState = actionEvaluator.PrepareInitialDelta(null);
             var evaluations = actionEvaluator.EvaluateTx(
                 blockHeader: block,
                 tx: tx,
@@ -564,7 +567,7 @@ namespace Libplanet.Tests.Action
                 Assert.Equal(
                     (Integer)eval.OutputState.GetState(
                         DumbAction.RandomRecordsAddress),
-                    (Integer)eval.InputContext.Random.Next());
+                    (Integer)eval.InputContext.GetRandom().Next());
                 IActionEvaluation prevEval = i > 0 ? evaluations[i - 1] : null;
                 Assert.Equal(
                     prevEval is null
@@ -594,7 +597,7 @@ namespace Libplanet.Tests.Action
 
             DumbAction.RehearsalRecords.Value =
                 ImmutableList<(Address, string)>.Empty;
-            previousState = actionEvaluator.PrepareInitialDelta(block);
+            previousState = actionEvaluator.PrepareInitialDelta(null);
             IAccount delta = actionEvaluator.EvaluateTx(
                 blockHeader: block,
                 tx: tx,
@@ -625,7 +628,7 @@ namespace Libplanet.Tests.Action
             var hash = new BlockHash(GetRandomBytes(BlockHash.Size));
             var actionEvaluator = new ActionEvaluator(
                 policyBlockActionGetter: _ => null,
-                blockChainStates: NullChainStates.Instance,
+                stateStore: new TrieStateStore(new MemoryKeyValueStore()),
                 actionTypeLoader: new SingleActionLoader(typeof(ThrowException))
             );
             var block = new BlockContent(
@@ -637,7 +640,7 @@ namespace Libplanet.Tests.Action
                     txHash: BlockContent.DeriveTxHash(txs),
                     lastCommit: CreateBlockCommit(hash, 122, 0)),
                 transactions: txs).Propose();
-            IAccount previousState = actionEvaluator.PrepareInitialDelta(block);
+            IAccount previousState = actionEvaluator.PrepareInitialDelta(null);
             var nextState = actionEvaluator.EvaluateTx(
                 blockHeader: block,
                 tx: tx,
@@ -768,7 +771,7 @@ namespace Libplanet.Tests.Action
             var block = chain.ProposeBlock(
                 GenesisProposer, txs.ToImmutableList(), CreateBlockCommit(chain.Tip));
 
-            IAccount previousState = actionEvaluator.PrepareInitialDelta(genesis);
+            IAccount previousState = actionEvaluator.PrepareInitialDelta(null);
             var evaluation = actionEvaluator.EvaluatePolicyBlockAction(genesis, previousState);
 
             Assert.Equal(chain.Policy.BlockAction, evaluation.Action);
@@ -787,7 +790,7 @@ namespace Libplanet.Tests.Action
             Assert.True(evaluation.InputContext.BlockAction);
 
             chain.Append(block, CreateBlockCommit(block), render: true);
-            previousState = actionEvaluator.PrepareInitialDelta(block);
+            previousState = actionEvaluator.PrepareInitialDelta(genesis.StateRootHash);
             var txEvaluations = actionEvaluator.EvaluateBlock(
                 block,
                 previousState).ToList();
@@ -980,7 +983,8 @@ namespace Libplanet.Tests.Action
             var miner = new PrivateKey();
             Block block = chain.ProposeBlock(miner);
 
-            var evaluations = chain.ActionEvaluator.Evaluate(block);
+            var evaluations = chain.ActionEvaluator.Evaluate(
+                block, chain.Store.GetStateRootHash(block.PreviousHash));
 
             Assert.False(evaluations[0].InputContext.BlockAction);
             Assert.Single(evaluations);
@@ -1043,7 +1047,9 @@ namespace Libplanet.Tests.Action
             var miner = new PrivateKey();
             Block block = chain.ProposeBlock(miner);
 
-            var evaluations = chain.ActionEvaluator.Evaluate(block);
+            var evaluations = chain.ActionEvaluator.Evaluate(
+                block,
+                chain.Store.GetStateRootHash(block.PreviousHash));
 
             Assert.False(evaluations[0].InputContext.BlockAction);
             Assert.Single(evaluations);
@@ -1110,7 +1116,8 @@ namespace Libplanet.Tests.Action
             var miner = new PrivateKey();
             Block block = chain.ProposeBlock(miner);
 
-            var evaluations = chain.ActionEvaluator.Evaluate(block);
+            var evaluations = chain.ActionEvaluator.Evaluate(
+                block, chain.Store.GetStateRootHash(block.PreviousHash));
 
             Assert.False(evaluations[0].InputContext.BlockAction);
             Assert.Single(evaluations);
@@ -1171,7 +1178,8 @@ namespace Libplanet.Tests.Action
             var miner = new PrivateKey();
             Block block = chain.ProposeBlock(miner);
 
-            var evaluations = chain.ActionEvaluator.Evaluate(block);
+            var evaluations = chain.ActionEvaluator.Evaluate(
+                block, chain.Store.GetStateRootHash(block.PreviousHash));
 
             Assert.False(evaluations[0].InputContext.BlockAction);
             Assert.Single(evaluations);
@@ -1295,7 +1303,7 @@ namespace Libplanet.Tests.Action
             {
                 IActionEvaluation eval = evalsA[i];
                 IActionContext context = eval.InputContext;
-                Assert.Equal(initialRandomSeed + i, context.Random.Seed);
+                Assert.Equal(initialRandomSeed + i, context.RandomSeed);
             }
         }